this should output english
20160712-00:27:47 mengwong@venice2:~/non-db-src/l/compiler/sandbox4% ghci mydeal.hs
GHCi, version 7.10.3: http://www.haskell.org/ghc/  :? for help
[1 of 2] Compiling L4v1             ( L4v1.hs, interpreted )
[2 of 2] Compiling Main             ( mydeal.hs, interpreted )
Ok, modules loaded: L4v1, Main.
*Main> main
This agreement dated Jun 30 2016
- Alice
- Bob
(the "Parties")

hereby agree as follows:

Effective immediately upon signature by all parties,

Alice is responsible for making a payment. Upon the effective date of this contract, Alice has 86400 seconds to pay Bob SGD 10. If this clause is performed, then the clause is fulfilled, but if it is not, then the contract is breached.

Bob is responsible for transferring an item. Upon the effective date of this contract, Bob has 86400 seconds to transfer Smoochy the Frog to Alice. If this clause is performed, then the clause is fulfilled, but if it is not, then the contract is breached.

*Main>
Leaving GHCi.