take command line arguments
20160727-21:17:06 mengwong@venice2:~/non-db-src/l/compiler/sandbox5-rounds/fundraising.hsproj% cd . ; runghc Fundraising.hs --years=8 --start=4000000 --end=5000000000
hello, world!
If initial valuation is 4,000,000 and want to get to 5,000,000,000 in 8 years,
then every year our valuation will need to increase by 2.4384496 times.

in 2016, we will be worth      4,000,000
in 2017, we will be worth      9,753,798
in 2018, we will be worth     23,784,146
in 2019, we will be worth     57,996,440
in 2020, we will be worth    141,421,408
in 2021, we will be worth    344,848,960
in 2022, we will be worth    840,896,832
in 2023, we will be worth  2,050,484,608
in 2024, we will be worth  5,000,003,584