refactor Gruter as L4 core. Particularize business logic
> import GruterIEng.gf
- compiling LexGruterEng.gf...   write file LexGruterEng.gfo
- compiling GruterIEng.gf...   write file GruterIEng.gfo
linking ... OK

Languages: GruterIEng

179 msec

2 msec
Gruter> gt | l -unlextext
By default, David Default may perform an action, where the action is correct.
By default, David Default shan't perform an action, where the action is correct.
By default, David Default shall perform an action, such that the action is correct.

0 msec
Gruter> e

0 msec
> import ParticularEng.gf
- compiling ParticularEng.gf...   write file ParticularEng.gfo
linking ... OK

Languages: ParticularEng
180 msec

Particular> gt -cat=Contract | l -unlextext
If the sun is shining, Alice Andromeda may have deliverred, where the things are hot.
If the sun is shining, Alice Andromeda may have payed, where the payment is correct.
If the sun is shining, Alice Andromeda shan't have deliverred, where the things are hot.
If the sun is shining, Alice Andromeda shan't have payed, where the payment is correct.
If the sun is shining, Alice Andromeda shall have deliverred, such that the things are hot.
If the sun is shining, Alice Andromeda shall have payed, such that the payment is correct.
If the sun is shining, Bai Bobo may have deliverred, where the things are hot.
If the sun is shining, Bai Bobo may have payed, where the payment is correct.
If the sun is shining, Bai Bobo shan't have deliverred, where the things are hot.
If the sun is shining, Bai Bobo shan't have payed, where the payment is correct.
If the sun is shining, Bai Bobo shall have deliverred, such that the things are hot.
If the sun is shining, Bai Bobo shall have payed, such that the payment is correct.
If the sun is shining, Carol Centaurus may have deliverred, where the things are hot.
If the sun is shining, Carol Centaurus may have payed, where the payment is correct.
If the sun is shining, Carol Centaurus shan't have deliverred, where the things are hot.
If the sun is shining, Carol Centaurus shan't have payed, where the payment is correct.
If the sun is shining, Carol Centaurus shall have deliverred, such that the things are hot.
If the sun is shining, Carol Centaurus shall have payed, such that the payment is correct.
When the moon is blue, Alice Andromeda may have deliverred, where the things are hot.
When the moon is blue, Alice Andromeda may have payed, where the payment is correct.
When the moon is blue, Alice Andromeda shan't have deliverred, where the things are hot.
When the moon is blue, Alice Andromeda shan't have payed, where the payment is correct.
When the moon is blue, Alice Andromeda shall have deliverred, such that the things are hot.
When the moon is blue, Alice Andromeda shall have payed, such that the payment is correct.
When the moon is blue, Bai Bobo may have deliverred, where the things are hot.
When the moon is blue, Bai Bobo may have payed, where the payment is correct.
When the moon is blue, Bai Bobo shan't have deliverred, where the things are hot.
When the moon is blue, Bai Bobo shan't have payed, where the payment is correct.
When the moon is blue, Bai Bobo shall have deliverred, such that the things are hot.
When the moon is blue, Bai Bobo shall have payed, such that the payment is correct.
When the moon is blue, Carol Centaurus may have deliverred, where the things are hot.
When the moon is blue, Carol Centaurus may have payed, where the payment is correct.
When the moon is blue, Carol Centaurus shan't have deliverred, where the things are hot.
When the moon is blue, Carol Centaurus shan't have payed, where the payment is correct.
When the moon is blue, Carol Centaurus shall have deliverred, such that the things are hot.
When the moon is blue, Carol Centaurus shall have payed, such that the payment is correct.

23 msec
Particular>